Move daemon spawning endpoints to `service._daemon` module
